squid:S2184 - Math operands should be cast before assignment

+++ src/org/jgroups/Message.java
@@ -864,7 +864,7 @@ public class Message implements Streamable {
     * @return The number of bytes for the marshalled message
     */
     public long size() {
-        long retval=Global.BYTE_SIZE   // leading byte
+        long retval=(long)Global.BYTE_SIZE   // leading byte
                 + Global.SHORT_SIZE;   // flags
         if(dest_addr != null)
             retval+=Util.size(dest_addr);
